apps/api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.32
alembic>=1.13.2
pydantic>=2.8.2
pydantic-settings>=2.4.0
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
resend>=2.3.0
